+++ samples/app/src/main/java/com/example/app/tweetui/ListTimelineFragment.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (C) 2015 Twitter, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.app.tweetui;
+
+import android.os.Bundle;
+import android.support.v4.app.ListFragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.app.R;
+
+import com.twitter.sdk.android.tweetui.TweetTimelineListAdapter;
+import com.twitter.sdk.android.tweetui.TwitterListTimeline;
+
+/**
+ * ListTimelineFragment demonstrates a TimelineListAdapter with a TwitterListTimeline.
+ */
+public class ListTimelineFragment extends ListFragment {
+
+    public static ListTimelineFragment newInstance() {
+        return new ListTimelineFragment();
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        final TwitterListTimeline timeline = new TwitterListTimeline.Builder()
+                .slugWithOwnerScreenName("twitter-bots", "dghubble")
+                .build();
+        final TweetTimelineListAdapter adapter = new TweetTimelineListAdapter(getActivity(),
+                timeline);
+        setListAdapter(adapter);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.tweetui_timeline, container, false);
+    }
+}