+++ tweet-ui/src/main/java/com/twitter/sdk/android/tweetui/UserTimeline.java
@@ -20,11 +20,8 @@ package com.twitter.sdk.android.tweetui;
 import com.twitter.sdk.android.core.Callback;
 import com.twitter.sdk.android.core.Result;
 import com.twitter.sdk.android.core.TwitterApiClient;
-import com.twitter.sdk.android.core.TwitterException;
 import com.twitter.sdk.android.core.internal.GuestCallback;
 import com.twitter.sdk.android.core.models.Tweet;
-
-import java.util.List;
 
 import io.fabric.sdk.android.Fabric;
 
@@ -89,38 +86,6 @@ public class UserTimeline extends BaseTimeline implements Timeline<Tweet> {
                         includeRetweets, new GuestCallback<>(new TweetsCallback(cb)));
             }
         };
-    }
-
-    /**
-     * Wrapper callback which unpacks a list of Tweets into a TimelineResult (cursor and items).
-     */
-    class TweetsCallback extends Callback<List<Tweet>> {
-        protected final Callback<TimelineResult<Tweet>> cb;
-
-        /**
-         * Constructs a TweetsCallback
-         * @param cb A callback which expects a TimelineResult
-         */
-        TweetsCallback(Callback<TimelineResult<Tweet>> cb) {
-            this.cb = cb;
-        }
-
-        @Override
-        public void success(Result<List<Tweet>> result) {
-            final List<Tweet> tweets = result.data;
-            final TimelineResult<Tweet> timelineResult
-                    = new TimelineResult<>(new TimelineCursor(tweets), tweets);
-            if (cb != null) {
-                cb.success(timelineResult, result.response);
-            }
-        }
-
-        @Override
-        public void failure(TwitterException exception) {
-            if (cb != null) {
-                cb.failure(exception);
-            }
-        }
     }
 
     /**